Adding suggest_template_info as the standard for suggestion UI rendering.
This proto serves as a template for rendering and fulfilling more
complex suggestions with subtitles and images (e.g. Entities and
Answers) along with their associated actions and
refinements/enhancement.

Note that entity-specific data, such as `entity_id` and `website_uri`
are encapsulated the existing EntityInfo and will be reported alongside
this template. Likewise, future work will handle Answer metadata like
attributions, which are not included in the current scope of this
template or this CL.

Change-Id: Ie4916b30645644c6995838d3ea1d6d1de8dedbe1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6524886
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Pierre St Juste <pstjuste@google.com>
Cr-Commit-Position: refs/heads/main@{#1458952}
NOKEYCHECK=True
GitOrigin-RevId: 93e699ed9e4bcbe6148fb33d0c5053f53c04f8f1