[canvas] Add shorthand headers for V8[Offscreen]RenderingContext
The autogenerated headers for the union types have enormous names. To
make it easier to update the union types, make forwarding headers
called that includes these giant filenames.

This should make it possible to add new options to the unions by
modifying the code in only 4 places (the union definition, the GN file,
the forwarding header, and the forward declaration in
CanvasRenderingContext which cannot use the autogenerated ones due to
include rules).

Bug: 413078308
Change-Id: I648dc87766db1ddf835088aa2e8771651913c418
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6501301
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459294}
NOKEYCHECK=True
GitOrigin-RevId: 8d387c21b507ec53caf76d78956a2d2a2067992c