Prompt API: Update MeasureInputUsage flow to use prompt
This CL makes the flow consistent with the writing assistance API where
the optional measurement result will be returned via a callback. The
client can now be removed.

It also makes some minor updates on the callback return type in the
measure methods in all the built-in AI APIs to returns uint32 instead of
uint64.

Bug: 351935691
Change-Id: Ic4a94f0310fad94bcb7b0be627ecd39526d491bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6529921
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Daseul Lee <dslee@chromium.org>
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459422}
NOKEYCHECK=True
GitOrigin-RevId: 8e04acc43a0c7f19f45025b48a60dce7e70af449