Ability to check inline/float/OOF without LayoutObject bits.
As pointed out in LayoutBox::IsValidColumnSpanner(), this function may
be called at "inconvenient" times, also before StyleDidChange() has
finished, so we cannot trust the LayoutObject bits for inline-level,
float, or position. Add the ability to provide this information without
checking those bits. The information we need is computed style and
layout object type.

It was possible for the function to return true for `column-span:all`
and `float:left`, which is wrong.

Bug: 417256483
Change-Id: Iaa8903a716a5f9a1d4cc18d42c48680e7fb2c551
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6540050
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459638}
NOKEYCHECK=True
GitOrigin-RevId: 518855753ca122d41906d52f9eb3bc389e1cadf5