De-share StyleSheetContents from the resource cache when needed
CSSStyleSheetResource caches its parsed StyleSheetContents, such that
multiple references to the same stylesheet (e.g. from various iframes
in the same origin) can share an instance of StyleSheetContents.
When mutating such stylesheets (e.g. insertRule), we must perform
a copy-on-write first, since they are supposed to be distinct stylesheets
behaviorally. However, the current flag we use to determine if
a StyleSheetContents is shared is not sufficient: we only check
if a given StyleSheetContents instance *currently* exists in the cache
(IsReferencedFromResource()), meaning that the flag can be *cleared*
when it's no longer the case. This means various CSSStyleSheet objects
which already hold a reference to a previously-cached StyleSheetContents
"forget" that it's shared.

To fix this, also set a flag whenever a StyleSheetContents is returned
as a cache hit.

This issue is believed to be the cause for the elusive
InspectorGhostRules crashes that has led to the feature being
reverted several times (CL:6529968).

Fixed: 417406834
Bug: 416699692, 417041745, 389011795
Change-Id: Idd340f7df2a2f1db0a634ea83798024aced12384
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6540832
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459687}
NOKEYCHECK=True
GitOrigin-RevId: d1bb719932208bcf428e7965c19e2bb5a50495d1