media: Fix MediaCapabilities reporting for audio only configs
In a previous CL http://crrev.com/c/6364074, we moved the KeySystem
check below audio and video support.

However, this introduced a bug because in audio-only configs, the code
previously had an optimization to return DecodeSupport early, without
checking KeySystemConfigs. With this code change, we now bring back
checking the key system support in the audio-only optimized path, and
introduce a browser test of audio only config with a key system
specified.

This now also introduces a KeySystemAccess check to verify that
KeySystemAccess object is not null when a valid KeySystemAccess is
provided.

Bug: 417251427
Change-Id: I0c29cc8e3597789b9ae85d936710f8e6c01af76c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6538246
Commit-Queue: Vikram Pasupathy <vpasupathy@chromium.org>
Reviewed-by: Sangbaek Park <sangbaekpark@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1460070}
NOKEYCHECK=True
GitOrigin-RevId: c06c66aad52c5d4d64e43a5b9fb9657eb224e43d