If an incoming break token is past block-end, so is the outgoing one.
This was failing for column balancing at a fragment that was for a
resumed node that had already overflowed, and had a forced break inside.
For this to happen, we need a balanced multicol container inside another
fragmentation context, and break in the outer fragmentation context,
then attempt to rebalance the inner while inside a resumed overflowed
node.

This fixes crashes, but simply avoiding those would leave a correctness
issue behind (since a break inside a parallel flow would have incorrect
impact on the outside). Hence a (rather complicated) correctness test
for this. Without this fix, the test would both fail a DCHECK and (if
bypassed) a CHECK.

Bug: 396176635
Change-Id: I08e4a6ca2832e027daf02f5a7e38e5b3c632a4ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6545960
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1460130}
NOKEYCHECK=True
GitOrigin-RevId: 79253bd274389cfa444381c0d6ca220b82cdc5f0