[vt] Remove obsolete and problematic CHECK.
Previously the destructor for the ViewTransitionStyleTracker ensured
that the state was idle or finished to ensure that the StyleEngine
did not retain a persistent list of active view transition names.
These names are now stored directly in the style tracker so the
CHECK is no longer required. In fact, the CHECK leaded to a crash
if the element that is the scope of view transition is removed and
garbage collected triggered. The style tracker is kept alive via
references from the pseudo-elements and view transition. The view
transition is in turn kept alive via a weak reference. Once the
element is removed, so are the pseudo-elements and the view
transition, resulting in the style tracker's destructor being called
while the state is not idle or finished.  A test has been added to
ensure we don't reintroduce a faulty assumption about the state of
the style tracker.

Bug: 415376109, 416811954
Change-Id: I91a1facd7f86f0c2c5a280348ea6416d34e6ed02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6544155
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Alexis Menard <alexis.menard@intel.com>
Cr-Commit-Position: refs/heads/main@{#1460265}
NOKEYCHECK=True
GitOrigin-RevId: b0a9da6636d9811b6f33d631aeac52195a9edb21