Replace 2 tests with a subtest of smoke/tentative/almost-soft-navigation.html.
replacestate.tentative.html is already covered with the replace-state
subtest, l. 59 in almost-soft-navigation.html in this change.

soft-navigation-no-url.tentative.html is getting covered with
noUrlPassedToPushState, in this change.

The tests I'm deleting make use of testSoftNavigationNotDetected in soft-navigation-helper.js, the replacements don't. Because the replacements don't wait for a timeout after each of these "almost softnavs" and run in the same doc and browser instance, they should be faster.

I'm also adding a global_test_id variable to make the
actual-softnavigation (which follows each attempted soft navigation)
go to a unique and predictable URL.

Change-Id: Id0d832540cc9293fbbeae2d99fce9a2e6d122a8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6545933
Commit-Queue: Johannes Henkel <johannes@chromium.org>
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1460312}
NOKEYCHECK=True
GitOrigin-RevId: 835ed1d8357d6797d527d3db3ae44111e462c3c6