Prototype anchored(fallback: <integer>)
Prototyping to inform explainer. The syntax is not decided.

This CL implements parsing of a new 'anchored' container-type along with
a new anchored() function that takes a 'fallback' feature matching the
chosen position-try-fallback with an 1-based index into the computed
position-try-fallback list. A value of 0 means no fallback chosen.

The evaluation is implemented, but CSSContainerValues always returns
that no fallback is chosen.

Bug: 417621241

Change-Id: I5b001ac644c0ee25384f8583c8abdfa603541b36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6544631
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1460357}
NOKEYCHECK=True
GitOrigin-RevId: c38a012d05222044111eb84ddf0052cce7940be0