HIC: Add tests for detached and display: none
We cannot draw into a detached canvas due to a lack of layout.
We cannot draw with display: none on the canvas or child due to
a lack of layout.

Also re-arrange the error checks to catch the most actionable thing
first and make the error messages as clear as possible.

Bug: 417556295
Change-Id: I5e3d5593594a1483fba0e909b1f0d3c0fc9ab4f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6542552
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1460417}
NOKEYCHECK=True
GitOrigin-RevId: 31e7e9fce83c0ae08dbd3492354f1a8e1bc15f4f