[A11yPerformance] Separate ax cache for ax tree snapshots
An existing AXObjectCache might not contain the nodes expected by
the snapshotter. For example, it might be in form controls mode or
onscreen-only mode. In this case, the tree snapshot could conflict
with the existing AT usage.

It's better to create a separate cache for the one-time snapshot,
to keep the states from ever having the possibility of a conflict.

This also helps land a performance refactoring for
AXObject::Serialize() in crrev.com/c/6509982, because that method
will now get a correct boolean state for whether something is for
a snapshot.

Potential follow-ups:
- Create a new class of cache, used for snapshots only, e.g.
  AXObjectCacheForTreeSnapshots, and don't allow either class of
  cache to be for the wrong purpose.
- The new cache class could potentially be simpler and do less
  work, although it may not be worth the effort to do this.

Bug: none
Change-Id: Iec0d6a21d08c4be544aa5cdfba60a79580d6c5c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6538187
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Matthias Körber <koerber@google.com>
Reviewed-by: Matthias Körber <koerber@google.com>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1460678}
NOKEYCHECK=True
GitOrigin-RevId: 53407869b1ae3d0646eb35e14ee39caba503de45