[gap-decorations] Avoid invalidating col rules when GapDecorations is on
This CL makes it so we avoid making a call to
`ColumnRuleStyleDidChange()` in `LayoutBlockFlow::StyleDidChange()` so
that we do not invalidate column rules when GapDecorations is enabled,
since with GapDecorations we already invalidate via the `invalidate:
paint` pipeline in `css_properties.json`.

The test added in this CL makes sure we are still invalidating correctly
after this change.

Bug: 357648037
Change-Id: I93f1a453133e658bca88277d8e253075ce2e2cd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6532734
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1460754}
NOKEYCHECK=True
GitOrigin-RevId: 0070897b2d63a31f4298ea8cd3dd035fefe08e6e