Use std::atomic to track in-flight speculative decode request
cc only ever designates one thread for running explicit image decode
requests, so we only issue one such request at a time from the main
thread. The downside to this is that the compositor thread must
PostTask to the main thread to signal that a decode request has
completed, and that task can be delayed if the main thread is busy,
causing us to miss an opportunity to get started on the next decode
ASAP.

The patch switches to using a std::atomic to signify whether a decode
is in-flight, allowing the main thread to detect sooner that the
next decode request can be issued.

Long-term, if speculative decode proves to be a valuable feature, it
probably makes sense to switch to using a thread-safe queue of decode
requests and integrate it into cc::Scheduler.

Bug: chromium:378623805
Change-Id: I1b882f37f36e86df8d6170d644b41cde411be070
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6533738
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1460805}
NOKEYCHECK=True
GitOrigin-RevId: 4dca3cd10b58c75c3ac6e09ab6216f2a12046ba6