Allow to mock multiple protocol in a single object
An Objective-C object can naturally implements an arbitrary number of
protocol. I ended up in a case where I had to modify my code, because
an object was expected to implement two protocols and OCMock does not
allow it. I’d love to see this code merged so that it can.

I also edited the code for nice mocks in order for the API to be
consistent. I don’t actually need it.

It should be noted that the description is not changed in the case
where there is a single protocol mocked.

Change-Id: I99f7c8f40041291d3eb1f8075334fa88bec60068
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6494931
Auto-Submit: Arthur Milchior <arthurmilchior@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1460973}
NOKEYCHECK=True
GitOrigin-RevId: bd5088ff3395df7935f7bee65304c429c004c256