[Blink] Fix BaseRC2D::getImageData() to not end hibernation
The call to create a CanvasResourceProvider here is a no-op beyond
having the unfortunate side-effect of waking up the canvas if it's
hibernating:

If we don't have that check and we can't create a resource provider in
GetImage() [1] we'll end up here [2] with a null snapshot, which means
that we'll be returning exactly the same thing we're currently returning
in the "short-circuit if resource provider can't be created" codepath.

This CL removes the call to create the CRP and adds a regression test
that hibernation is preserved through a call to getImageData().

Note: This necessitates explicitly handling the case where GetImage()
returns null to ensure that the image that will be returned from
getImageData() is zero-initialized. Prior to this CL that case was (I
believe) not hit in practice because the cases in which GetImage()
would return null are effectively the same cases in which
GetOrCreateCanvasResourceProvider() would cause getImageDataInternal()
to short-circuit out earlier.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/modules/canvas/canvas2d/canvas_rendering_context_2d.cc;l=658-660;drc=26fa5b6a22b36be60bc17cccd5754c44dcff3334
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/modules/canvas/canvas2d/base_rendering_context_2d.cc;l=585-587;drc=26fa5b6a22b36be60bc17cccd5754c44dcff3334

Bug: 40280152
Change-Id: I74096a85681de7c4d73dab4b357cb6c1055dbf7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6446396
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1461187}
NOKEYCHECK=True
GitOrigin-RevId: ed9e63e5a8b8bf4f8d009ef1f564336dfd51f0ab