[owners] Remove skyostil@chromium.org from third_party/blink/renderer/platform/instrumentation/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/skyostil@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ib273d482edd41fd596db92edb528ba4853a46163
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6552792
Reviewed-by: Mikhail Khokhlov <khokhlov@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>
Cr-Commit-Position: refs/heads/main@{#1461246}
NOKEYCHECK=True
GitOrigin-RevId: c50679b100602e10134b026b1be5ebcc161dfb36