[owners] Remove mheikal@chromium.org from third_party/android_build_tools/jetifier/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and approve
it unless you know a reason the OWNER should remain. It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/mheikal@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Id952d6bb52e06a9dcc87a79e3582767706a4554d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6557986
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1461477}
NOKEYCHECK=True
GitOrigin-RevId: 223c2440defb41dc86c3f91d789bdfe9a6b2be62