[owners] Remove mheikal@chromium.org from third_party/android_deps/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and approve
it unless you know a reason the OWNER should remain. It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/mheikal@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I1d329e9f1cf79ed17b2b6c982f6c0751a25921e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6556641
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1461492}
NOKEYCHECK=True
GitOrigin-RevId: b6d06c29fcae590e79577a592c46ae7e2346264e