[owners] Remove mheikal@chromium.org from third_party/android_build_tools/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and approve
it unless you know a reason the OWNER should remain. It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/mheikal@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ia05622311f5b334dc93626fc980b42df71765b1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6558047
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1461557}
NOKEYCHECK=True
GitOrigin-RevId: 930a3c358bfc56dd04ff9c115c5f7d7906d5d520