Closing interest invoker target popover loses interest
Since it's possible that the target popover for an interest
invoker (via `interesttarget`) can be closed through other means
such as light dismiss or programmatic `.hidePopover()`, that
should cause a loss of interest. Without this, the UI can get
into an odd state where the target popover is already closed,
yet the invoker still "has interest". To get the popover to
show up again, the invoker needs to be de-hovered *and left
de-hovered for the "hide" delay* so that interest can be lost.
Since that's completely invisible, the user has no way to know
whether it's "done" and the invoker can be re-hovered. With
this change, the UX will be intuitive.

Bug: 364669918
Change-Id: Ifab53af643ffdccfda47524b187255c756b3a106
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6536349
Commit-Queue: Mason Freed <masonf@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1461575}
NOKEYCHECK=True
GitOrigin-RevId: 916f2f6b54b510f383d60df79db8701ec72bba9a