Update WebXR Depth-Sensing WPT coverage
Consume new changes for depth-sensing added to the webxr-test-api. This
allows increased testing of `matchDepthView` and raw/smooth depth type
capabilities. Further test changes to e.g. cpu/gpu and supported formats
could now be made, but are left as an exercise for a future change, as
this change is primarily about increasing coverage for the most recent
spec additions that are intended to ship soon.

Note that due to the need to add new test asserts, the asserts file was
quite significantly re-worked. However, as no new test failures are
introduced, these changes should be a no-op that simply cleans up the
code a bit.

There is also an update to the webxr-samples to allow testing the
toggling of depthActive with dom-overlay.

Finally, GPU test failure expectations were updated. As chrome does not
support `gpu-optimized` depth, these tests should simply be all-up
disabled on all platforms, and don't need to be careful of specific
error messages.

Changes to webxr-test-api that enable this change can be found at:
https://github.com/immersive-web/webxr-test-api/pull/88

Bug: 410607163,409806803
Change-Id: I2dfdc2376a310b084fee75d521c712d5ae788433
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6554023
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1461626}
NOKEYCHECK=True
GitOrigin-RevId: f1c77012e716e27f4469a3f5217270b931f3afbc