[AdTracker] Don't constantly scan dynamically added inline scripts
The code would scan the top of the stack and insert the inline script
into the set of known ads if any current ad script was running. This
doesn't really make sense, we should only make a decision on whether
script is ad script once, when it is first introduced (e.g., top-level
execution).

This CL restricts scanning of dynamically inlined scripts (those with an
id but no script url) to top-level execution only and verifies the
correct behavior with a test.

Bug: 417950837
Change-Id: I900c6634555040654d9e0f0a32133101d7a160e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6558350
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Josh Karlin <jkarlin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1461720}
NOKEYCHECK=True
GitOrigin-RevId: fd853645fadb2a89dcc01bfc668e2ebe69b744e7