[owners] Remove bingler@chromium.org from third_party/blink/web_tests/external/wpt/cookies/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and approve
it unless you know a reason the OWNER should remain. It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/bingler@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I2753d0781374d1654c0cb8627701d6968447765f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6556771
Commit-Queue: Steven Bingler <bingler@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Steven Bingler <bingler@chromium.org>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Cr-Commit-Position: refs/heads/main@{#1461896}
NOKEYCHECK=True
GitOrigin-RevId: 341e143ffe9f0484269b4a217d5db4089cf95f6c