Add a LRU cache for relative URLs resolved in Document
This CL adds a LRU cache for relative URLs resolved in Document to
reduce URL parsing and copies. Document::CompleteURL is updated with the
memoization, and the cache is cleared when the base URL is updated to a
different URL, as the resolved URLs will become invalid.

Bug: 348442535
Change-Id: I8716df87dc27c6685da73bebf56ab6e2893c39b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6534472
Commit-Queue: Nidhi Jaju <nidhijaju@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1461907}
NOKEYCHECK=True
GitOrigin-RevId: c0d7b032d706e0b2096ff3e1bb262349d0bfe800