Setup WebAuthenticationImmediateGet for origin trial
Followed //docs/origin_trials_integration.md to set it up.

While the feature is set as enabled, the runtime guards will disable
it on the blink side unless the chrome://flags entry is set to enabled.

The feature state is set back to default, as features with experimental
or test status will report immediateGet in client capabilities. We
should update the platform tests to include immediateGet once the specs
are updated.

Bug: 393055190
Change-Id: I3b82c1bcde64634bfdd65d4ad671a3ba9f1cf1a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6536508
Commit-Queue: Adem Derinel <derinel@google.com>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Stephen White <senorblanco@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1461992}
NOKEYCHECK=True
GitOrigin-RevId: 8601e788b27a98d982ba015075307adc81b13f01