Fix ch fallback logic
This CL fixes the logic error introduced in crrev.com/c/6333369,

In the cases where it is impossible or impractical to determine the
measure of the “0” glyph, it must be assumed to be 0.5em wide by 1em
tall. Thus, the ch unit falls back to 0.5em in the general case,
and to 1em when it would be typeset upright (i.e. writing-mode is
vertical-rl or vertical-lr and text-orientation is upright).[1]

[1]https://drafts.csswg.org/css-values-4/#ch

Bug: 416145497
Change-Id: I5000ede2edfe02a4ea0df71e6511078d6de56651
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6520498
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Jason Leo <cgqaq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1462035}
NOKEYCHECK=True
GitOrigin-RevId: 9b7d041aae94972dbc0fd8091b8203dc8cad7809