[owners] Remove claudiomagni@chromium.org from third_party/blink/public/strings/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/claudiomagni@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ia0cd96aa3bb106f512fc8078f6853afe50f43ef0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6560679
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1462187}
NOKEYCHECK=True
GitOrigin-RevId: 5d65b2401906a90014ccfe101cd6c45919061e57