[flex-balance] Allow clamping of margin-box size to zero.
For the score based flex line-breaker to work (fast) we need the scoring
function to be monotonic. E.g. if we add an item to a line, the score
must decrease.

Negative margins break this invariant. There are various strategies to
mitigate this. For the moment we just clamp the margin-box size to zero
when this occurs.

An alternate more complex strategy is to perform a forward-scan, and
accumulate these negative-size items into the current items size. We'd
still need clamping in this case, as the negative margins could exceed
the current items size for example.

Lets do the simple thing for the moment, and see what feedback we get.

This patch introduces a `greedy_line_count` lambda to line-break off the
prefix-sums array, as we'll need this for the min-line-count feature
(and the prefix-sums array contains the source of truth of the sizes
now).

Bug: 416755656
Change-Id: I2bb5b8280c015d73ab8365b32bdb2f7852a830d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6557623
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1462202}
NOKEYCHECK=True
GitOrigin-RevId: 12cc431ee4a3180312e4fb13f5b54bb2b71be71d