[owners] Remove bingler@chromium.org from third_party/blink/public/common/storage_key/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and approve
it unless you know a reason the OWNER should remain. It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/bingler@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Id1455b2d2c2b84072d9fdebd5571c04bdfe9b6f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6557288
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Steven Bingler <bingler@chromium.org>
Commit-Queue: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1462223}
NOKEYCHECK=True
GitOrigin-RevId: 039bc1b0a2747ba2d7fdbc0e10b98233ad330a7e