[owners] Remove marcheu@chromium.org from third_party/libdrm/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/marcheu@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Icadf3f7e41cabc09728b452147b32815208f5669
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6554164
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1462229}
NOKEYCHECK=True
GitOrigin-RevId: 43bb260654cf2930f07b505874af69ca664cb94e