[owners] Remove chasej@chromium.org from third_party/blink/common/origin_trials/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/chasej@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I1697d288a7431604d6a7c25e1946c33e7777dc9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6554909
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1462270}
NOKEYCHECK=True
GitOrigin-RevId: 380f821dbdba0a31988905a28268a8b640d3151d