[owners] Remove adetaylor@chromium.org from third_party/domato/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/adetaylor@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Iedd57df6702a8c2aa245ed8e0c1dcb8a98c039d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6561043
Commit-Queue: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Paul Semel <paulsemel@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1462302}
NOKEYCHECK=True
GitOrigin-RevId: 1d927e8def654ac86e5572eeada248f101a14101