[owners] Remove jrummell@chromium.org from third_party/widevine/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and approve
it unless you know a reason the OWNER should remain. It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/jrummell@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ib6c77814d9642b986c3eeceaea4b6eb55309776f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6556672
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1462310}
NOKEYCHECK=True
GitOrigin-RevId: 4bcbdaeca1c7ac08f8618d4328a89f258025b861