[owners] Remove kyleju@chromium.org from third_party/wpt_tools/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/kyleju@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ia4c7c1c8544cc6d599ee4f400111f1c13b10bd4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6565739
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Commit-Queue: Weizhong Xia <weizhong@google.com>
Cr-Commit-Position: refs/heads/main@{#1462362}
NOKEYCHECK=True
GitOrigin-RevId: b5131e04239c7ef87b674b1a6ad3e511f687fed3