Make Windows touch drag and drop open context menu on static drags
This CL ties the touch_dragend_context_menu pref to the touch drag and
drop setting in Windows.
In Windows we want to open the context menu when a touch drag and drop
operation is released in the same place it started.

Bug: 411164103
Change-Id: Ie4e9cb1d0d1b346578d92dc8f867839a8987fe1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6496615
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Reviewed-by: Olga Gerchikov <gerchiko@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1462445}
NOKEYCHECK=True
GitOrigin-RevId: b3f013164b74f4b6bc778a444d24339fd4877029