[gap-decorations] Rule color is transparent fix for repeater values
This CL fixes an issue where we will crash when recomputing
the column-rule-color style when it is set to a repeater value.

In order to fix this, we consider the rule transparent on the case
of repeater values only if every individual value is transparent.

Bug: 357648037
Change-Id: I7072e7e4035af7ddad3acd26a01406d50219d410
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6536229
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Sam Davis Omekara <samomekarajr@microsoft.com>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1462485}
NOKEYCHECK=True
GitOrigin-RevId: 0273b508c87276e2b39f56af3385ba02c36ac5ac