Revert WPT changes to tiled-gradients.html.
It is not reasonable to demand that every pixel in a hard line comes
at exactly the same spot when repeated, which the fuzz changes to this
WPT test now does (breaking it in Chromium on every platform).
It also changed indentation for no good reason (breaking the blame
history) and removed the trailing newline.

Thus, revert the change.

Fixed: 418466380
Change-Id: I392ec5e36eea8719c3974691f672898c5922d45c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6565623
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1462684}
NOKEYCHECK=True
GitOrigin-RevId: f8ec5755a453325fefd1fedde22cafdcea75056d