Replace BoxBorderPainter Path argument with "side type"
A bunch of the functions in BoxBorderPainter take a Path argument of one
form or another. After previous cleanups it's however clear that it is
not actually used as a Path, but only as a glorified (and expensive)
bool.

Replace the Path argument to PaintOneBorderSide() with a "side type"
enumeration instead and then drop the unused Path arguments.

This shows that BoxBorderPainter::ComplexBorderInfo::rounded_border_path
is not used as well, so drop that.

Rename the functions with a ...FromPath() suffix to something else.

Bug: 418931564
Change-Id: Ia7da7d87d201724f4d17df29357f74a2804a0768
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6564416
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1462730}
NOKEYCHECK=True
GitOrigin-RevId: 2fd5775f141fd1a87e5b0f736d10affe58d3a711