[flex-balance] Implement min-line-count feature.
This implements the "balance in into a least N lines feature".

This does this by bisecting the line-break size - this is a
"high-constant" complexity, e.g. bisecting will finished worst case
31 iterations of the loop (position of MSB in reality).

This may not reach the desired number of lines however. E.g. if you
have a 2x2 grid of items there is no line-break size which will
achieve 3 lines.

To fix this we search for items which are "perfectly fitting", e.g.
items which with greedy line-breaking at against the very end of the
line - and increase them by 1 ScoreUnit.

This forces them onto a subsequent line, increasing the total number
of lines (and doesn't materially affect scoring).

Bug: 416755656
Change-Id: Iad8e477d4ddb4899d3ca6baa3bcbc5aaeac5d7a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6546460
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1462858}
NOKEYCHECK=True
GitOrigin-RevId: 9539a4f602822719c1b46971c98b7a22eab0f384