[flex-balance] During measure divide the available-size by min-lines.
Typically when measuring, the available-size is the content-size as we
don't know how many lines we are going to end up with.

However if balancing with a minimum number of lines it yields better
results if we divide the available-size by the min-line count.

Bug: 416755656
Change-Id: Ib32e00767eacd340918717ff55b84a9fefe6913e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6565876
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1462860}
NOKEYCHECK=True
GitOrigin-RevId: 6a2798b9fc4d93a3d2978efbe1ef67e63cc95509