corner-shape: support outline
When using an outline, use the border path's stroke
as the outline clip path.

This technique is possible in outlines and not in borders, because the outline thickness cannot change between sides.

Using round stroke caps/joins to avoid very long miters.
This is discussed in the CSSWG
(https://github.com/w3c/csswg-drafts/issues/12194)
and can be amended if something else comes up, however currently outline
rendering is not specified in an interoperable way, so adding the tests
as regular pixel web-tests.

Note that dashed/dotted outlines are still not rendered correctly, so
not adding tests for them. The same goes for auto outlines, and for
complex outlines that span multiple boxes.

Bug: 417420173
Change-Id: I16bbae1b1ea3e5a29a33b6e60ae6a8c381082e4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6551430
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Commit-Queue: Noam Rosenthal <nrosenthal@google.com>
Cr-Commit-Position: refs/heads/main@{#1462900}
NOKEYCHECK=True
GitOrigin-RevId: d5b55732e21428a9d102779b544ccbb466876adb