[WebNN] Fuse QDQ for softmax on tflite
This CL supports fusing `dq->softmax->q` subgraph on tflite.

Input and output operands have to be dequantized from ints8, the scale
and zero point of input and output have to be scaler. For TFLite kernel,
the scale of output should be approximately equal to 1.0f / 256.0f and
the zero point of output should be equal to -128 if data type is int8.

This CL also fixes the wrong output tensor's shape of softmax.

Bug: 401281047, 338094927
Change-Id: Ie7ae32d3b71f7f396552622da3b8195192ba19b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6565957
Commit-Queue: Wei4 Wang <wei4.wang@intel.com>
Reviewed-by: ningxin hu <ningxin.hu@intel.com>
Reviewed-by: Phillis Tang <phillis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1463148}
NOKEYCHECK=True
GitOrigin-RevId: df243e84bfcf8735a471a5a7da1ab81e6b746e45