[rebaseline-cl] Parallelize allowed suffix check
https://crrev.com/c/5011405 added a check so that `-expected.png` are
never copied for reftests. However, checking whether a legacy web test
is a reftest is slow, so computing this up-front in the manager process
is a bottleneck. Defer this check into the workers for parallelism.

This CL has no functional changes, but reduces the time to pool start of
`optimize-baselines --all` from ~45s to <6s.

Bug: 411245983
Change-Id: I136988a993c7d4e871c3b4282ffd97aae7c47b82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6569591
Commit-Queue: Jonathan Lee <jonathanjlee@google.com>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Cr-Commit-Position: refs/heads/main@{#1463149}
NOKEYCHECK=True
GitOrigin-RevId: b993cb7c35fa24208ccdd6ad85a9e034c8945b61